manager no longer remote
I commented out the list comp and used map for 1 cpu evals...I know in diff cases one may be faster than the other but hard to say in general according to random stack overflow discussions. I left both in(lines 54/55) to better decide when integrating to Deap official in case Deap Team has a preference, conform to them.

There was no good reason to run the manager as a remote worker via ray. This only added overhead and in the case of num_cpus=1 it really hurt where as standard map should be allowed to work on the main process.

Also removed the unneeded creator and optional pset setups on the manager, adding them to scope was no longer needed once the manager was on the main process(see above).